move Iterators module out of Base
