+++ base/client.jl
@@ -46,7 +46,7 @@ const disable_text_style = AnyDict(
 
 # Create a docstring with an automatically generated list
 # of colors.
-available_text_colors = collect(Iterators.filter(x -> !isa(x, Integer), keys(text_colors)))
+available_text_colors = collect(Iter.Filter(x -> !isa(x, Integer), keys(text_colors)))
 const possible_formatting_symbols = [:normal, :bold, :default]
 available_text_colors = cat(1,
     sort!(intersect(available_text_colors, possible_formatting_symbols), rev=true),
