+++ base/precompile.jl
@@ -607,14 +607,14 @@ precompile(Tuple{typeof(Base.setindex!), Base.Dict{Union{DataType, typeof(Type)}
 precompile(Tuple{typeof(Base.setindex!), Base.Dict{Union{DataType, typeof(Type)}, Nothing}, Nothing, typeof(Type)})
 precompile(Tuple{typeof(Base.rehash!), Base.Dict{Union{DataType, typeof(Type)}, Nothing}, Int64})
 precompile(Tuple{typeof(Base.isempty), Array{Int64, 1}})
-precompile(Tuple{typeof(Base.Iterators.zip), Array{Int64, 1}, SimpleVector})
-precompile(Tuple{typeof(Base.Iterators.zip), Array{Symbol, 1}, SimpleVector})
-precompile(Tuple{typeof(Base.start), Base.Iterators.Zip2{Array{Int64, 1}, SimpleVector}})
-precompile(Tuple{typeof(Base.start), Base.Iterators.Zip2{Array{Symbol, 1}, SimpleVector}})
-precompile(Tuple{typeof(Base.done), Base.Iterators.Zip2{Array{Int64, 1}, SimpleVector}, Tuple{Int64, Int64}})
-precompile(Tuple{typeof(Base.done), Base.Iterators.Zip2{Array{Symbol, 1}, SimpleVector}, Tuple{Int64, Int64}})
-precompile(Tuple{typeof(Base.next), Base.Iterators.Zip2{Array{Int64, 1}, SimpleVector}, Tuple{Int64, Int64}})
-precompile(Tuple{typeof(Base.next), Base.Iterators.Zip2{Array{Symbol, 1}, SimpleVector}, Tuple{Int64, Int64}})
+precompile(Tuple{typeof(Base.Iter.zip), Array{Int64, 1}, SimpleVector})
+precompile(Tuple{typeof(Base.Iter.zip), Array{Symbol, 1}, SimpleVector})
+precompile(Tuple{typeof(Base.start), Base.Iter.Zip2{Array{Int64, 1}, SimpleVector}})
+precompile(Tuple{typeof(Base.start), Base.Iter.Zip2{Array{Symbol, 1}, SimpleVector}})
+precompile(Tuple{typeof(Base.done), Base.Iter.Zip2{Array{Int64, 1}, SimpleVector}, Tuple{Int64, Int64}})
+precompile(Tuple{typeof(Base.done), Base.Iter.Zip2{Array{Symbol, 1}, SimpleVector}, Tuple{Int64, Int64}})
+precompile(Tuple{typeof(Base.next), Base.Iter.Zip2{Array{Int64, 1}, SimpleVector}, Tuple{Int64, Int64}})
+precompile(Tuple{typeof(Base.next), Base.Iter.Zip2{Array{Symbol, 1}, SimpleVector}, Tuple{Int64, Int64}})
 precompile(Tuple{typeof(Base.println), Base.GenericIOBuffer{Array{UInt8, 1}}, DataType})
 precompile(Tuple{typeof(Base.println), Base.GenericIOBuffer{Array{UInt8, 1}}, typeof(Type)})
 precompile(Tuple{typeof(Base.setindex!), Base.Dict{Any, Any}, Array{Base.Docs.DocStr, 1}, Symbol})
