+++ base/reflection.jl
@@ -48,8 +48,8 @@ Get the fully-qualified name of a module as a tuple of symbols. For example,
 
 # Examples
 ```jldoctest
-julia> fullname(Base.Iterators)
-(:Base, :Iterators)
+julia> fullname(Base.Sys)
+(:Base, :Sys)
 
 julia> fullname(Main)
 (:Main,)
