+++ doc/make.jl
@@ -102,7 +102,6 @@ const PAGES = [
         "base/io-network.md",
         "base/punctuation.md",
         "base/sort.md",
-        "base/iterators.md",
         "base/c.md",
         "base/libc.md",
         "base/stacktraces.md",
