+++ stdlib/Iterators/docs/src/index.md
@@ -1,0 +1,17 @@
+# Iteration utilities
+
+```@docs
+Iterators.zip
+Iterators.enumerate
+Iterators.rest
+Iterators.countfrom
+Iterators.take
+Iterators.drop
+Iterators.cycle
+Iterators.repeated
+Iterators.product
+Iterators.flatten
+Iterators.partition
+Iterators.filter
+Iterators.reverse
+```
