+++ test/reflection.jl
@@ -88,7 +88,7 @@ i10165(::Type{AbstractArray{T,n}}) where {T,n} = 1
 
 # fullname
 @test fullname(Base) == (:Base,)
-@test fullname(Base.Iterators) == (:Base, :Iterators)
+@test fullname(Base.Libc) == (:Base, :Libc)
 
 const a_const = 1
 not_const = 1
